Send SIGCONT after detaching
Just wait for a second after detaching, that should be enough for the
kernel to do it's thing, then send the SIGCONT.

This might cancel a legitimate SIGSTOP that didn't come from us, but I
don't care. If you're throwing a SIGSTOP at the game in the time after
sekirofpsunlock detaches but before it sends the SIGCONT, you should be
equipped to deal with this kind of shit, anyway.

I just can't figure out why the game *sometimes* remains stopped after
detaching. One thing that I'm sure of is that ptrace is incredibly
complicated; we deserve a better syscall.